+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implement `SourceIter` and `InplaceIterable` for `IntoIter` when
+  `unstable_nightly` is enabled.
+
+### Changed
+- Synchronize source code with standard library of Rust version 1.48.0.
 
 ## [0.2.0] &ndash; 2020-10-26
 ### Changed